Private Documentation

Documentation for GridPotentialFlow.jl's private interface.

See the Internals section of the manual for internal package docs covering all submodules.

Contents

Private Documentation

Contents

Index

Private Interface

Index

Base.collect

GridPotentialFlow.setposition!

GridPotentialFlow.setpositions!

Private Interface

Base.collect — Method

collect(vl::VortexList) -> Tuple{Array{Float64,1},Array{Float64,1},Array{Float64,1}}

Collect the inertial-space coordinates and strengths of all of the Lagrange points comprising the vortices in vortex list vl and return each assembled set as a vector.

GridPotentialFlow.setposition! — Method

setposition!(vortex::Vortex, xnew::Real, ynew::Real) -> Vortex

Sets the x and y fields of vortex to the xnew and ynew, respectively.

GridPotentialFlow.setpositions! — Method

setpositions!(vortices::VortexList, xpositions::Any, ypositions::Any) -> Any

Sets the x and y fields of the point vortices in vortices to the entries of xpositions and ypositions, respectively.