Public Documentation

Documentation for GridPotentialFlow.jl's public interface.

See the Internals section of the manual for internal package docs covering all submodules.

Contents

Public Documentation

Contents

Index

Public Interface

Index

GridPotentialFlow.ModelParameters

GridPotentialFlow.ModelParameters

GridPotentialFlow.Vortex

GridPotentialFlow.VortexList

GridPotentialFlow.VortexModel

GridPotentialFlow.VortexModel

GridPotentialFlow.computeaddedmassmatrix

GridPotentialFlow.computeimpulse

GridPotentialFlow.computeimpulse

GridPotentialFlow.computevortexvelocities

GridPotentialFlow.computevortexvelocities

GridPotentialFlow.computew

GridPotentialFlow.computew!

GridPotentialFlow.computeψ

GridPotentialFlow.getpositions

GridPotentialFlow.getstrengths

GridPotentialFlow.getvortexpositions

GridPotentialFlow.pushvortices!

GridPotentialFlow.setvortexpositions!

GridPotentialFlow.setvortices!

GridPotentialFlow.solvesystem

GridPotentialFlow.solvesystem!

GridPotentialFlow.updateposition!

Public Interface

GridPotentialFlow.ModelParameters — Type

struct ModelParameters

Defines the model parameters for VortexModel.

Fields

Ub::Union{Nothing, AbstractArray}

Ub: Array that contains the velocity of each body or nothing if there is no body. Each entry should be a collection of the x- and y-components of the velocity.

U∞::Any

U∞: Collection of the x- and y-components of the free stream velocity.

Γb::Union{Nothing, AbstractArray}

Γb: Array that contains the desired circulation of each body or nothing if there is no body. This will only be enforced if the body has no regularized edges.

σ::Union{Nothing, AbstractArray{Union{Float64, SuctionParameterRange},N} where N}

σ: Array that contains the suction parameter or suction parameter range for each edge or nothing if there are no regularized edges or if the Kutta condition is to be enforced.

GridPotentialFlow.ModelParameters — Method

ModelParameters(; Ub, U∞, Γb, σ) -> ModelParameters

Constructs the parameters for a vortex model using the given function.

GridPotentialFlow.Vortex — Type

mutable struct Vortex

Defines a point vortex with x-position x, y-position y, and strength Γ.

Fields

x::Float64

x: x-coordinate of the vortex position.

y::Float64

y: y-coordinate of the vortex position.

Γ::Float64

Γ: strength of the vortex. Positive if counter-clockwise.

GridPotentialFlow.VortexList — Type

struct VortexList

Defines a list of point vortices.

Fields

list::Array{Vortex,1}

list: vector of point vortices.

GridPotentialFlow.VortexModel — Type

mutable struct VortexModel{Nb, Ne, isshedding}

Defines a grid-based vortex model with Nb bodies and Ne edges that are regularized. If isshedding is true, the strengths of the last Ne vortices in vortices can be computed in order to regularized the Ne edges.

Fields

g::PhysicalGrid

g: The grid on which the vortex model is defined.

bodies::BodyList

bodies: The bodies in the vortex model.

vortices::VortexList

vortices: The point vortices in the vortex model.

edges::Array{Int64,1}

edges: The vector of body points where shedding occurs.

system::PotentialFlowSystem

system: The potential flow system that has to be solved with a PotentialFlowRHS and a PotentialFlowSolution to compute the potential flow that governs the vortex model.

_nodedata::Nodes{Dual,NX,NY,T,DT} where DT<:(AbstractArray{T,2} where T) where T<:Number where NY where NX

Internal fields

_edgedata::Edges{Primal,NX,NY,T,DT} where DT where T<:Number where NY where NX

_bodydata::ScalarData

_bodyvectordata::VectorData

_d::Nodes{Dual,NX,NY,T,DT} where DT<:(AbstractArray{T,2} where T) where T<:Number where NY where NX

_d_kvec::Array{Nodes{Dual,NX,NY,T,DT} where DT<:(AbstractArray{T,2} where T) where T<:Number where NY where NX,1}

_ψ::Nodes{Dual,NX,NY,T,DT} where DT<:(AbstractArray{T,2} where T) where T<:Number where NY where NX

_f::ScalarData

_w::Nodes{Dual,NX,NY,T,DT} where DT<:(AbstractArray{T,2} where T) where T<:Number where NY where NX

_ψb::ScalarData

Examples

(under construction)

GridPotentialFlow.VortexModel — Method

VortexModel(g::PhysicalGrid; bodies, vortices, edges) -> VortexModel{_A,_B,_C} where _C where _B where _A

Constructs a vortex model using the given function.

GridPotentialFlow.computeaddedmassmatrix — Method

Computes the translational coefficients of the added mass matrix of the bodies in vortexmodel.

GridPotentialFlow.computeimpulse — Method

computeimpulse(vortexmodel; parameters)

Computes the impulse associated with the current state vortices and bodies in vortexmodel.

GridPotentialFlow.computeimpulse — Method

computeimpulse(vortexmodel, wphysical, fphysical, Ubvec)

Computes the impulse associated with the vorticity wphysical, the bound vortex sheet strength fphysical, and the velocities Ubvec of the discrete points of the bodies in vortexmodel.

GridPotentialFlow.computevortexvelocities — Method

computevortexvelocities(vortexmodel, ψ)

Returns the flow velocity as VectorData at the locations of the vortices stored in vortexmodel associated with the discrete vector potential field ψ.

GridPotentialFlow.computevortexvelocities — Method

Returns the flow velocity as VectorData at the locations of the vortices stored in vortexmodel, accounting for bodies in vortexmodel and conditions in kwargs. If the vortexmodel has Ne regularized edges and vortices, the strengths of the last Ne vortices will be computed and set in vortexmodel.

GridPotentialFlow.computew! — Method

computew!(wphysical, vortexmodel)

Computes the vorticity field w associated with the vortices stored in vortexmodel on the physical grid.

GridPotentialFlow.computew — Method

Computes the vorticity field associated with the vortices stored in vortexmodel on the physical grid.

GridPotentialFlow.computeψ — Method

computeψ(vortexmodel; kwargs...)

Computes and returns the stream function field on the physical grid for the potential flow associated with the current state of vortexmodel. A uniform flow, body velocities, bound circulation values for the unregularized bodies, and suction parameters for regularized bodies can be specified using the parameters keyword.

GridPotentialFlow.getpositions — Method

getpositions(vortices::VortexList) -> VectorData{_A,Float64,Array{Float64,1}} where _A

Returns the positions of all point vortices in vortices as VectorData.

GridPotentialFlow.getstrengths — Method

getstrengths(vortices::VortexList) -> ScalarData{_A,Float64,Array{Float64,1}} where _A

Returns the strengths of all point vortices in vortices as ScalarData.

GridPotentialFlow.getvortexpositions — Method

Returns the positions of the vortices in vortexmodel.

GridPotentialFlow.pushvortices! — Method

pushvortices!(vortexmodel, newvortices)

Adds the newvortices to the existing vortices in vortexmodel.

GridPotentialFlow.setvortexpositions! — Method

setvortexpositions!(vortexmodel, X_vortices)

Sets the positions of the vortices in vortexmodel to the provided X_vortices.

GridPotentialFlow.setvortices! — Method

setvortices!(vortexmodel)
setvortices!(vortexmodel, newvortices)

Replaces the vortices field of vortexmodel by a copy of newvortices.

GridPotentialFlow.solvesystem! — Method

solvesystem!(sol, vortexmodel, wphysical; parameters)

Computes the potential flow solution sol for the bodies in vortexmodel and vorticity on the physical grid wphysical. If the vortexmodel contains bodies with regularized edges and a number of vortices which is greater than or equal to the number of regularized edges, the returned solution contains the computed strengths of the last N vortices in vortexmodel, with N the number of regularized edges. A uniform flow, body velocities, bound circulation values for the unregularized bodies, and suction parameters for regularized bodies can be specified using the parameters keyword.

GridPotentialFlow.solvesystem — Method

solvesystem(vortexmodel, wphysical; kwargs...)

Computes and returns the potential flow solution for the bodies in vortexmodel and vorticity on the physical grid wphysical. If the vortexmodel contains bodies with regularized edges and a number of vortices which is greater than or equal to the number of regularized edges, the returned solution contains the computed strengths of the last N vortices in vortexmodel, with N the number of regularized edges. A uniform flow, body velocities, bound circulation values for the unregularized bodies, and suction parameters for regularized bodies can be specified using the parameters keyword.

GridPotentialFlow.updateposition! — Method

updateposition!(vortex::Vortex, u::Real, v::Real, Δt::Real) -> Vortex

Updates the x and y fields of vortex, given the x-velocity u, y-velocity v, and timestep Δt.